@@ -3,21 +3,21 @@
   import TodosList from './TodosList.svelte'
   import EditTodo from './EditTodo.svelte'
   import AddTodo from './AddTodo.svelte'
-  import { views, editingIndex } from '../lib/stores'
+  import { views, editingIndex } from '../lib/stores.svelte'
 
   const queryClient = useQueryClient()
 </script>
 
 <div>
   <div>
-    <button on:click={() => queryClient.invalidateQueries()}>
+    <button onclick={() => queryClient.invalidateQueries()}>
       Force Refetch All
     </button>
   </div>
   <br />
   <hr />
 
-  {#each $views as view}
+  {#each views.value as view}
     <div>
       <TodosList initialFilter={view} />
       <br />
@@ -25,15 +25,15 @@
   {/each}
 
   <button
-    on:click={() => {
-      views.set([...$views, ''])
+    onclick={() => {
+      views.value.push('')
     }}
   >
     Add Filter List
   </button>
   <hr />
 
-  {#if $editingIndex !== null}
+  {#if editingIndex.value !== null}
     <EditTodo />
     <hr />
   {/if}
