@@ -2,14 +2,13 @@
   import '../app.css'
   import { QueryClientProvider } from '@tanstack/svelte-query'
   import { SvelteQueryDevtools } from '@tanstack/svelte-query-devtools'
-  import type { PageData } from './$types'
 
-  export let data: PageData
+  const { data, children } = $props()
 </script>
 
 <QueryClientProvider client={data.queryClient}>
   <main>
-    <slot />
+    {@render children()}
   </main>
   <SvelteQueryDevtools />
 </QueryClientProvider>
