@@ -2,24 +2,24 @@
   import { createQuery } from '@tanstack/svelte-query'
   import { getPlanet } from '$lib/api'
 
-  export let homeworldId: string
+  let { homeworldId }: { homeworldId: string } = $props()
 
-  const query = createQuery({
+  const query = createQuery(() => ({
     queryKey: ['homeworld', homeworldId],
     queryFn: () => getPlanet(homeworldId),
-  })
+  }))
 </script>
 
-{#if $query.status === 'pending'}
+{#if query.status === 'pending'}
   <span>Loading...</span>
 {/if}
 
-{#if $query.status === 'error'}
+{#if query.status === 'error'}
   <span>Error :(</span>
 {/if}
 
-{#if $query.status === 'success'}
+{#if query.status === 'success'}
   <span>
-    {$query.data.name}
+    {query.data.name}
   </span>
 {/if}
