@@ -1,16 +1,27 @@
 <script lang="ts">
   import { useHydrate } from './useHydrate.js'
+  import type { Snippet } from 'svelte'
   import type {
     DehydratedState,
     HydrateOptions,
     QueryClient,
   } from '@tanstack/query-core'
 
-  export let state: DehydratedState
-  export let options: HydrateOptions | undefined = undefined
-  export let queryClient: QueryClient | undefined = undefined
+  type Props = {
+    children: Snippet
+    state: DehydratedState
+    options: HydrateOptions | undefined
+    queryClient: QueryClient | undefined
+  }
+
+  const {
+    children,
+    state,
+    options = undefined,
+    queryClient = undefined,
+  }: Props = $props()
 
   useHydrate(state, options, queryClient)
 </script>
 
-<slot />
+{@render children()}
