@@ -2,8 +2,10 @@
   import { onDestroy, onMount } from 'svelte'
   import { QueryClient } from '@tanstack/query-core'
   import { setQueryClientContext } from './context.js'
+  import type { QueryClientProviderProps } from './types.js'
 
-  export let client = new QueryClient()
+  const { client = new QueryClient(), children }: QueryClientProviderProps =
+    $props()
 
   onMount(() => {
     client.mount()
@@ -16,4 +18,4 @@
   })
 </script>
 
-<slot />
+{@render children()}
