@@ -1,12 +1,9 @@
 <script lang="ts">
-  import { QueryClient } from '@tanstack/query-core'
   import { QueryClientProvider } from '../../src/index.js'
   import ChildComponent from './ChildComponent.svelte'
-  import type { QueryCache } from '@tanstack/query-core'
+  import type { QueryClient } from '@tanstack/query-core'
 
-  export let queryCache: QueryCache
-
-  const queryClient = new QueryClient({ queryCache })
+  let { queryClient }: { queryClient: QueryClient } = $props()
 </script>
 
 <QueryClientProvider client={queryClient}>
