@@ -3,25 +3,25 @@
   import { createInfiniteQuery } from '../../src/index.js'
   import { sleep } from '@tanstack/query-test-utils'
 
-  export let queryClient: QueryClient
+  let { queryClient }: { queryClient: QueryClient } = $props()
 
   const queryKey = ['test']
 
-  let firstPage = 0
+  let firstPage = $state(0)
 
   const query = createInfiniteQuery(
-    {
+    () => ({
       queryKey: queryKey,
       queryFn: ({ pageParam }) => sleep(10).then(() => pageParam),
       getNextPageParam: (lastPage) => lastPage + 1,
       initialPageParam: firstPage,
-    },
-    queryClient,
+    }),
+    () => queryClient,
   )
 </script>
 
 <button
-  on:click={() => {
+  onclick={() => {
     queryClient.setQueryData(queryKey, {
       pages: [7, 8],
       pageParams: [7, 8],
@@ -32,4 +32,4 @@
   setPages
 </button>
 
-<div>Data: {JSON.stringify($query.data)}</div>
+<div>Data: {JSON.stringify(query.data)}</div>
