@@ -1,23 +1,23 @@
 <script lang="ts">
-  import { writable } from 'svelte/store'
   import { QueryClient } from '@tanstack/query-core'
   import { createMutation, setQueryClientContext } from '../../src/index.js'
 
-  export let mutationFn: (value: {
-    count: number
-  }) => Promise<{ count: number }>
+  let {
+    mutationFn,
+  }: { mutationFn: (value: { count: number }) => Promise<{ count: number }> } =
+    $props()
 
-  const count = writable(0)
+  let count = $state(0)
 
   const queryClient = new QueryClient()
   setQueryClientContext(queryClient)
 
-  const mutation = createMutation({ mutationFn })
+  const mutation = createMutation(() => ({ mutationFn }))
 </script>
 
-<button on:click={() => $mutation.mutate({ count: ++$count })}>Mutate</button>
+<button onclick={() => mutation.mutate({ count: ++count })}>Mutate</button>
 
-<div>Data: {$mutation.data?.count ?? 'undefined'}</div>
-<div>Status: {$mutation.status}</div>
-<div>Failure Count: {$mutation.failureCount}</div>
-<div>Failure Reason: {$mutation.failureReason ?? 'undefined'}</div>
+<div>Data: {mutation.data?.count ?? 'undefined'}</div>
+<div>Status: {mutation.status}</div>
+<div>Failure Count: {mutation.failureCount}</div>
+<div>Failure Reason: {mutation.failureReason ?? 'undefined'}</div>
