@@ -6,13 +6,13 @@
   const queryClient = new QueryClient()
   setQueryClientContext(queryClient)
 
-  const mutation = createMutation({
+  const mutation = createMutation(() => ({
     mutationFn: () =>
       sleep(10).then(() => Promise.reject(new Error('Expected mock error'))),
-  })
+  }))
 </script>
 
-<button on:click={() => $mutation.reset()}>Reset</button>
-<button on:click={() => $mutation.mutate()}>Mutate</button>
+<button onclick={() => mutation.reset()}>Reset</button>
+<button onclick={() => mutation.mutate()}>Mutate</button>
 
-<div>Error: {$mutation.error?.message ?? 'undefined'}</div>
+<div>Error: {mutation.error?.message ?? 'undefined'}</div>
