@@ -9,32 +9,32 @@
     forks_count: number
   }
 
-  const query = createQuery<Repo>({
+  const query = createQuery<Repo>(() => ({
     queryKey: ['repoData'],
     queryFn: async () =>
       await fetch('https://api.github.com/repos/TanStack/query').then((r) =>
         r.json(),
       ),
-  })
+  }))
 </script>
 
 <h1>Simple</h1>
 <div class="my-4">
   <div>
-    {#if $query.isPending}
+    {#if query.isPending}
       Loading...
     {/if}
-    {#if $query.error}
+    {#if query.error}
       An error has occurred:
-      {$query.error.message}
+      {query.error.message}
     {/if}
-    {#if $query.isSuccess}
+    {#if query.isSuccess}
       <div>
-        <h1>{$query.data.full_name}</h1>
-        <p>{$query.data.description}</p>
-        <strong>👀 {$query.data.subscribers_count}</strong>{' '}
-        <strong>✨ {$query.data.stargazers_count}</strong>{' '}
-        <strong>🍴 {$query.data.forks_count}</strong>
+        <h1>{query.data.full_name}</h1>
+        <p>{query.data.description}</p>
+        <strong>👀 {query.data.subscribers_count}</strong>{' '}
+        <strong>✨ {query.data.stargazers_count}</strong>{' '}
+        <strong>🍴 {query.data.forks_count}</strong>
       </div>
     {/if}
   </div>
