@@ -1,3 +1,4 @@
+import type { Snippet } from 'svelte'
 import type {
   DefaultError,
   DefinedQueryObserverResult,
@@ -11,14 +12,13 @@ import type {
   MutationState,
   OmitKeyof,
   Override,
+  QueryClient,
   QueryKey,
   QueryObserverOptions,
   QueryObserverResult,
 } from '@tanstack/query-core'
-import type { Readable } from 'svelte/store'
 
-/** Allows a type to be either the base object or a store of that object */
-export type StoreOrVal<T> = T | Readable<T>
+export type Accessor<T> = () => T
 
 /** Options for createBaseQuery */
 export type CreateBaseQueryOptions<
@@ -33,7 +33,7 @@ export type CreateBaseQueryOptions<
 export type CreateBaseQueryResult<
   TData = unknown,
   TError = DefaultError,
-> = Readable<QueryObserverResult<TData, TError>>
+> = QueryObserverResult<TData, TError>
 
 /** Options for createQuery */
 export type CreateQueryOptions<
@@ -68,13 +68,13 @@ export type CreateInfiniteQueryOptions<
 export type CreateInfiniteQueryResult<
   TData = unknown,
   TError = DefaultError,
-> = Readable<InfiniteQueryObserverResult<TData, TError>>
+> = InfiniteQueryObserverResult<TData, TError>
 
 /** Options for createBaseQuery with initialData */
 export type DefinedCreateBaseQueryResult<
   TData = unknown,
   TError = DefaultError,
-> = Readable<DefinedQueryObserverResult<TData, TError>>
+> = DefinedQueryObserverResult<TData, TError>
 
 /** Options for createQuery with initialData */
 export type DefinedCreateQueryResult<
@@ -134,9 +134,7 @@ export type CreateMutationResult<
   TError = DefaultError,
   TVariables = unknown,
   TOnMutateResult = unknown,
-> = Readable<
-  CreateBaseMutationResult<TData, TError, TVariables, TOnMutateResult>
->
+> = CreateBaseMutationResult<TData, TError, TVariables, TOnMutateResult>
 
 /** Options for useMutationState */
 export type MutationStateOptions<TResult = MutationState> = {
@@ -144,4 +142,9 @@ export type MutationStateOptions<TResult = MutationState> = {
   select?: (
     mutation: Mutation<unknown, DefaultError, unknown, unknown>,
   ) => TResult
+}
+
+export type QueryClientProviderProps = {
+  client: QueryClient
+  children: Snippet
 }
