@@ -1,6 +1,6 @@
 import { getIsRestoringContext } from './context.js'
-import type { Readable } from 'svelte/store'
+import type { Box } from './containers.svelte.js'
 
-export function useIsRestoring(): Readable<boolean> {
+export function useIsRestoring(): Box<boolean> {
   return getIsRestoringContext()
 }
