@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
 import { render } from '@testing-library/svelte'
-import { QueryCache } from '@tanstack/query-core'
+import { QueryClient } from '@tanstack/query-core'
 import ParentComponent from './ParentComponent.svelte'
 
 describe('QueryClientProvider', () => {
@@ -13,11 +13,12 @@ describe('QueryClientProvider', () => {
   })
 
   test('Sets a specific cache for all queries to use', async () => {
-    const queryCache = new QueryCache()
+    const queryClient = new QueryClient()
+    const queryCache = queryClient.getQueryCache()
 
     const rendered = render(ParentComponent, {
       props: {
-        queryCache: queryCache,
+        queryClient: queryClient,
       },
     })
 
