@@ -1,11 +1,10 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
-import { QueryClient } from '@tanstack/query-core'
 import { fireEvent, render } from '@testing-library/svelte'
-import { get, writable } from 'svelte/store'
+import { QueryClient } from '@tanstack/query-core'
+import { ref } from '../utils.svelte.js'
 import BaseExample from './BaseExample.svelte'
 import SelectExample from './SelectExample.svelte'
 import ChangeClient from './ChangeClient.svelte'
-import type { Writable } from 'svelte/store'
 import type { QueryObserverResult } from '@tanstack/query-core'
 
 describe('createInfiniteQuery', () => {
@@ -18,21 +17,20 @@ describe('createInfiniteQuery', () => {
   })
 
   it('should return the correct states for a successful query', async () => {
-    const statesStore: Writable<Array<QueryObserverResult>> = writable([])
+    let states = ref<Array<QueryObserverResult>>([])
 
     const rendered = render(BaseExample, {
       props: {
-        states: statesStore,
+        states,
       },
     })
 
     await vi.advanceTimersByTimeAsync(11)
     expect(rendered.getByText('Status: success')).toBeInTheDocument()
 
-    const states = get(statesStore)
+    expect(states.value).toHaveLength(2)
 
-    expect(states).toHaveLength(2)
-    expect(states[0]).toEqual({
+    expect(states.value[0]).toEqual({
       data: undefined,
       dataUpdatedAt: 0,
       error: null,
@@ -68,7 +66,8 @@ describe('createInfiniteQuery', () => {
       fetchStatus: 'fetching',
       promise: expect.any(Promise),
     })
-    expect(states[1]).toEqual({
+
+    expect(states.value[1]).toEqual({
       data: { pages: [0], pageParams: [0] },
       dataUpdatedAt: expect.any(Number),
       error: null,
@@ -107,25 +106,25 @@ describe('createInfiniteQuery', () => {
   })
 
   it('should be able to select a part of the data', async () => {
-    const statesStore: Writable<Array<QueryObserverResult>> = writable([])
+    let states = ref<Array<QueryObserverResult>>([])
 
     const rendered = render(SelectExample, {
       props: {
-        states: statesStore,
+        states,
       },
     })
 
     await vi.advanceTimersByTimeAsync(11)
     expect(rendered.getByText('count: 1')).toBeInTheDocument()
 
-    const states = get(statesStore)
+    expect(states.value).toHaveLength(2)
 
-    expect(states).toHaveLength(2)
-    expect(states[0]).toMatchObject({
+    expect(states.value[0]).toMatchObject({
       data: undefined,
       isSuccess: false,
     })
-    expect(states[1]).toMatchObject({
+
+    expect(states.value[1]).toMatchObject({
       data: { pages: ['count: 1'] },
       isSuccess: true,
     })
