@@ -1,5 +1,4 @@
 import { describe, expectTypeOf, test } from 'vitest'
-import { get } from 'svelte/store'
 import {
   QueriesObserver,
   QueryClient,
@@ -46,11 +45,11 @@ describe('queryOptions', () => {
       queryFn: () => Promise.resolve(5),
     })
 
-    const queries = createQueries({
+    const queries = createQueries(() => ({
       queries: [options],
-    })
+    }))
 
-    expectTypeOf(get(queries)[0].data).toEqualTypeOf<number | undefined>()
+    expectTypeOf(queries[0].data).toEqualTypeOf<number | undefined>()
   })
 
   test('Should tag the queryKey with the result type of the QueryFn', () => {
